fix - ofMesh.load() resolves an issue where Normals would not be loaded back in.
When saving a .ply file using ofMesh::save() normals are stored as vertex element properties. These were, however, not imported in back again in ofMesh.load() because ofMesh.load() expected normals to be in an element list.

This fix brings ofMesh::load() in line with ofMesh::save() and the .ply file format specs by loading normals as vertex element attributes if normals are declared in the .ply.

See test case here:

http://github.com/tgfrerer/testMeshLoader
Signed-off-by: Tim Gfrerer <tim.gfrerer@oo43.com>