fixes built-in shaders
there is a massive weirdness going on in the bitmapstringVertexShader. it needs a placeholder attribute, aptly named 'placeholder' if it is not present, the shader will not render correctly, and glDrawArrays will get you an error. after hours of hunting this down, i am prone to think that this could be an ATI on mac os x driver bug.

I'll check it on linux.

Debugging raw openGL calls on os x is a bloody nightmare.

Signed-off-by: tgfrerer <tim@poniesandlight.co.uk>