add example for instanced vboMesh drawing
This demonstrates how to use ofVboMesh::drawInstanced() using the programmable GL renderer and a matching shader.

It will draw 16384 instances of an ofBox primitive, spreading these out on a grid, extruding their height on the vertex shader using a depth texture, and animating along the x-axis, to make it a little more interesting.

Plenty of comments inline.

Signed-off-by: tgfrerer <tim@poniesandlight.co.uk>