fix ofLight cone debug draw if spotlight
this makes sure the spotlight cone points in the right direction when a spotlight ofLight is drawn using the fixed function renderer.

because the cone primitive has its base circle in x/z and the tip on the y-axis, we need to rotate the cone first, so that it points along the z-axis.

Signed-off-by: tgfrerer <tim@poniesandlight.co.uk>