GPU optim # 1
store uniform locations even if location returns -1. If uniform location would not be stored, many unnecessary glGetUniformLocation requests would be issued on every draw call.

This is especially pronounced if custom shaders were used which do not define the built-in shader uniforms.

Since the uniform locations cannot change once a shader is linked, a failure to look up the uniform on the first frame will also mean a failure to look it up on every following frame. By catching this early, we don't incur a driver overhead penalty per uniform per frame.