Update ofUtils.cpp
Changed `char c;` to `int c;` because fgetc() returns int and EOF is also int. Otherwise the comparison `c != EOF` fails in my tests on Raspberry Pi (Raspbian Wheezy) and ofSystem() gets stuck in the do-while-loop. (That's probably because EOF is defined as -1 and when fgetc returns this to a char variable, it is saved as 255 and a comparison with EOF fails.)