Update ofShader.cpp to use logging modules
There were a number of logging messages produced by ofShader that did not have the module set to "ofShader". This update fixes most of these. However, due to limitations in the ofLog interface there is no easy way to add module information to the ofLog messages that take a log level as an argument in the checkShaderInfoLog() function. To do so would need a constructor for ofLog that takes a module string as an argument or some other mechanism for providing module information.