fix ofShader load return value
Currently, `ofShader::load()` loads, compiles, and links GLSL shader source.

`ofShader::load()` returns a bool, which is set based on the current shader object's `bLoaded` member variable, which does not say anything about whether the shader has been loaded, compiled and linked sucessfully.

It will also attempt to compile and link shader stages, even if an earlier shader stage compile or load failed.

---

This PR suggests the following changes:

* If an error is encountered in any of the stages of ofShader::load (`load`, `compile`, and `link`), return `false` early at the point of error, print an error messge, and do not continue processing this shader. This makes it easier to debug shader errors, since subsequent errors only fill up the log, and are most likely just results of the first error reported.

* ofShader::linkProgram only returns true if *both* load and `linkStatus`  were true, that is, when the shader was successfully linked.

With this, the return value of `ofShader::load()` is a reliable indicator of whether the shader was loaded, compiled, and linked properly.