fix state leak when recursively drawing MSAA fbos
When resolving MSAA fbos, global fbo binding state must be
changed, because the resolve is a blit from one fbo into
another.

This PR makes sure to correctly keep track of the global
renderer FBO binding state and to restore it after the
blitting operation has completed.

+ Also includes some minor typo fixes for the doxygen
documentation of the updateTexture() method.