fix vao not updated upon vertex buffer change
When an ofVbo is assigned a new buffer for vertex data, the
internal VAO was not flagged as dirty. This would lead to
the old buffer still being bound to the previous vertex
buffer, and the new vertex data not being displayed.

With this fix, the ofVbo's internal VAO is marked dirty
after the buffer has been assigned, and the VAO will be
re-submitted upon the next draw call, using the new buffer
for vertex data.