Fixes osx ofToDataPath. See https://github.com/openframeworks/projectGenerator/pull/22
In osx when OF starts it sets the current directory to the directory
where the exe is located and then uses a relative path from there to
correctly find the data folder.

This has several problems:

- The current working directory is not correct anymore, mostly for command
line applications is useful to know where the command was run from. This is
for example making the command line PG fail in osx and would require a hack
like https://github.com/openframeworks/projectGenerator/pull/22, in general
people would need to know that OF internally changes the cwd but even with
that the actual cwd is totally lost so there's no way to recover it except
by queryng $PWD which is not multiplatform

- If someone changes the current working directory using setcwd or std::current_path
the data folder won't be found anymore

Instead of that this commit changes the data path to EXE_DIR/../../..data
which always works no matter if the cwd is changed by the application or
where the application is run.

This also changes `ofFilePath::getCurrentWorkingDir` which right now is returning
the exe dir instead of the cwd