ofParameter: add ofParameter<void>
which it's pretty useless by itself but allows to add a parameter to
an ofParameterGroup that will end up represented as a button in an
ofxGui.

an ofParameter<void> can't be set or have operators... but has a
`trigger()` method that calls the corresponding event.

This still doesn't add the psoibility to add an ofParameter<void> to
an ofxPanel.