fix arealights for programmable renderer
because the old implementation mixed the STRINGIFY
preprocessor macro with GLSL preprocessor directives, the
behaviour of the c preprocessor for this file was undefined.

This lead to weird GLSL code in VS2015 builds.

To close #4495, this PR suggests & implements the following:

* Using Raw String Literals instead of STRINGIFY allows us to
  concat different cpp constants based on the cpp build
  environment. Raw String Literals are a C++11 feature. See:
  http://en.cppreference.com/w/cpp/language/string_literal

* Anything within Raw String Literals is not interpreted
  by the preprocessor, also escapes are not scanned. It is
  literally a literal.

Tested on XCode 7.1 / clang 7.0.0 and VS2015; Raw string
literals are advertised as supported by GCC from 4.5
onwards. https://gcc.gnu.org/projects/cxx0x.html