fix scrolling behaviour for ofxSlider
This updates ofxSlider to use the new .scrollY property of
ofMouseEventArgs, and also fixes a possible bug where the
scroll event would have fired even if the cursor was not
above the active element.

With .scrollY and .y both being in place, the gui element
may now find out correctly, whether the cursor is above it
whilst the scroll event fires.