Addresses 4 issues related to #4556:
The macro ofxNetworkCheckError() and the function it calls ofxNetworkCheckErrno() make no attempt to convert parameters to strings avoiding calling any function that might reset the errno or (WSA)GetLastError() value.
ofxNetworkCheckErrno() will not generate a log entry if (WSA)EWOULDBLOCK is the error.
ofxTCPManager::Receive() checks the error returned from ofxNetworkCheckError() and sets the return value to 0 (instead of -1) if (WSA)EWOULDBLOCK is returned, indicating no data available and not an error.
On some OSs EAGAIN might be defined as a different value from EWOULDBLOCK.