Update changelog for 0.9.1
This is not ready! I just wanted to start a discussion - if I should continue summarizing the commits, and putting the issue numbers up, or if I should just copy-paste the commit log in question. I am trying to summarize to make the changelog more human readable, but maybe instead of linking to the github issue number, I should link to the git log commit hashes?