use clang/libc++ for android
c++11 suport in libstdc++ in android is not complete and fails to compile some
libraries that use such features. This changes the android toolchaing to use
clang + libc++ instead of gcc + libstdc++ and adapts all the apothecary formulas
to use that same toolchain too.

This doesn't include the opencv libraries yet since opencv requires version 3.1 to
be able to compile with libc++ so that will go in a different PR