Fix for ply mesh data loader with uchar color type
Color vertec attributes may be specified as uchar, not
float. The existing implementation was not taking this into
consideration, and interpreting uchars as float.

This PR suggests adding a flag for the vertex color type
derived by reading the file's meta section.

It then makes interpretation of color data dependent on this
flag, and scales uchar color data to normalised float range.

Fixes #4935