fix ofNode destructor when parented
+ safer destruction of ofNode

ofNodes with parents/children would not properly destroy,
since the destructor in ofNode would stumble over an
invalidated iterator since the children's list of a node
would get altered when deleting a child's parent's
children list, which points back, don't ask a genealogist,
to the object itself.

+ clear parent in setParent if parenthood is re-assigned

when re-assigning parenthood to an object that already had
a parent, listeners assigned by the previous parenthood were
not cleared. This should be fixed  now.

+ clear up parameter ambiguity in ofNode::setParent

To increase readability, the parameter in the member
function referring to the parent to be set is renamed.