ofShader: add methods to change constant and defines values
in a shader with a define like:

you can now call:

shader.setDefineConstant(USE_SOME_FEATURE, 0);

to disable it.

There's a similar method for const numerical or boolean values. in
this case the types are more strict similar to how setUniform* works.

This recompiles the shader so it's slow to use but it can be faster
than uniforms in several cases, usually when it's used for values
that won't change often and would introduce a conditional or a for
loop with bounds comming from a uniform which makes the compiler
unable to unroll the loop.