ofNode: remove node from mat4
decomposing a mat4 to it's components: position, orientation and scale
is not very reliable, mostly if it's not known if the matrix might contain
both scale and rotation since both are stored in the matrix diagonal

this removes the method that created a node from a mat4 since it would
mostly work but not always was really confusing for people using it
thinking it was completely reliable.

as a side effect setParent and clearParent with the maintain global
transform option on are now recreated from the individual components
being much more robust.