fix ofPolyLine::calcData returning invalid values
This fixes a regression where ofPolyline would return unnormalized values
for tangent, rotation and normal.

The issue existed, because the glm:: implementation of glm::normalize
don't operate in-place, but their return value must be used.

Addresses issue #5288

+ fixes an issue where pathological polylines would return nan vectors

Whenever a calculation is done which involves normalisation we should
check for the length of the vector being normalised, or the
normalization result not being NaN (which is the result of normalizing a
0-vector in glm). oF used to return a zero vector in this case, and this
implementation, based on the orignal code from rev:0361224e tries to
match the expected result.

Also, when two segments are congruent i.e. they point exactly against
each other, this implementation will use the direction of one segment as
tangent, so that a meaningful normal can be calculated. This is an edge
case, and it should probably discussed how best to implement this.

+ also updates comments in .h file