Settings for EGLWindow were getting lost.
When setup(const ofWindowSettings &) was called, it was expected
that the inherited setup(const ofWindowSettings &) method would be called.
But it was not.

Instead, a temporary of type ofGLESWindowSettings was being created
and setup(const ofGLESWindowSettings &) was being called.
When the temporary was created, the EGL settings were lost.

By adding a "using" statement, the proper method is called.