External time mode
Adds a new time mode, external, that allows to set the global OF
time calls to return the time based on an external source.

A common case could be a sound source for synchronizing animations,
to sound.

This includes a new funcion in fmod: ofFmodGetDSPTime that can be used
with this new functionality as:

```
ofSetTimeModeExternal(ofFmodGetDSPTime, ofTime());
```

although for sound is recomended to use:

```
ofSetTimeModeExternalFiltered(ofFmodGetDSPTime, ofTime(), 0.99);
```

or similar since the different update rates can make the ofGetLastFrameTime
very noisy otherwise