vs: update download_libs.ps1
+ adds vs_ver as a selectable option to the powershell download script,
which allows you to select which version of the libraries to download.

Default paramter value for vs_ver is 2015, which will load the libraries
compiled with the visual studio 2015 (== 140) toollset.

Call the script with:

    powershell.exe -f download_libs.ps1 -vs_ver 2017

To download & install the libraries compiled with the visual studio 2017
(== 141) toolset.

This should also fix an issue where the old download_libs.ps1 script
would not work, as the libraries archives on the server now have a
slightly different naming scheme.