fix ofCamera::drawFrustum
+ `drawFrustum` would foolishly assume ofRectMode to be set to its
default value, `OF_RECTMODE_CENTER` when drawing the frustum caps. This
might lead to frustum caps being drawn incorrectly when, somewhere
before, rectmode had been set to something other than
`OF_RECTMODE_CORNER`.  This is fixed now - style is pushed before
drawing the frustum, set correctly for drawing the frustum, and restored
to user specified values after drawing the frustum, using
ofPush/PopStyle.

+ also adds parameter to optionally explicitly define a viewport from
which to calculate the frustum to be drawn - this brings the method more
in line with other camera-specific methods, which allow optional
specification of a viewport.

By default, the current viewport state is used, as before.