fix download_libs.ps1 script
+ script was deleting temporary libs folder "32", "64" repeatedly
meaning only the libs from the very last .zip slice of libs would
survive the cull.

Since the temporary folders "32/64" get deleted automatically as they
are part of the "cull list" when preparing the libs directory for
upgrade this additional delete routine is not needed in the first place.

+ also added an echo to instruct users on how to install libs
for visual studio 2017, instead of the default, visual studio 2015.

+ also added an echo telling users that that the script finished
successfully.