fix teardown order
+ makes sure app is destroyed before corresponding window is closed

when a window was closed using GUI controls any OS-native method of
closing a window, window and app teardown did not follow the same order
as it would have if the window+app would have been closed if you had hit
the ESC key.

This brings both teardown control paths back in line, and is the first
step to fix an isssue with teardown which may happen in OpenGL Windows
apps.