force destruction of renderer before window
+ if the last application window is destroyed, this may in some cases
cause the OpenGL library to be unloaded. If the renderer was still alive
at this time, any OpenGL calls would call errors on teardown (whilst
closing an app).

This change makes sure that the renderer for a window is destroyed
before the corresponding window, so that any leftover OpenGL calls have
finished executing before the OpenGL context may be closed when
destroying the window.