ofPath: flagShapeChanged() and hasChanged() now marked public
ofPath::flagShapeChanged() could be really useful when paths are changed externally (the workaround is to call ofPath::getCommands() only so that ofPath::flagShapeChanged is called), and ofPath::hasChanged doesn't get called at all anywhere in OF, so making it public doesn't hurt the internals. Also, making it public mirrors the behavior of ofMesh::hasChanged.