Fix for ofAVFoundationGrabber.mm setDevice(...) after close()
Fix for a bug outlined in this forum post : https://forum.openframeworks.cc/t/changing-the-camera-input-while-the-app-is-running/7622/5

If you attempt to set the device ID after you have called close() on the grabber, the grabber object no longer exists, and any subsequent calls to setup() will use the default device (id=0). I have added a check that reallocates the OSXVideoGrabber if necessary.