Fixes for html5 audio and video with emscripten (#6039)
SoundPlayer.isLoaded() was giving a wrong output on html5 audio. It returned if the loading started, not if the loading was finished. This is caused by the way the audio is fetched (asynchronous).

Video grabbing from webcam was outdated and did not work on IOS11 and Safari 11.x.

This fixes problems, updates the deprecated navigator.getUserMedia() to navigator.mediaDevices.getUserMedia and includes a fix to get webcam support on both IOS and Safari.