fix bounding box calculations, letter placing
letterSpacing controls the default kerning between letters in a relative
manner. Since proportional fonts have per-letter advance values, we use
the current width of a space character of the font as the base for
letterSpacing.

This means, letterSpacing of 1.0 means neutral letter spacing, glyph
is placed in default position, respecting the glyph's advance value.

    + If letterSpacing is 0.75, this that spacing between letters is
    reduced by 0.25 * the whitespace width of the current font.

    + If letterSpacing is 1.25, this means that an additional space of
    0.25 * the advance of glyph 'p' is added between letters.

+ additionally, per spec, we scale the default whitespace glyph width by
the font's property `spaceSize`

Note that to calculate the default whitespace width for a font we don't
use the advance value for the literal whitespace (' ') glyph, but, as
the docs in the comments specify, we take the whitespace width as the
advance of the font's glyph 'p'.