ofxGui: slider scrolling (#6144)
Fixes #6133

The scrolling was still being handled when the gui was not being drawn both in ofxSlider and ofxInputField.

Tested using the code provided in the mentioned issue.

Also,
there is some redundant code in ofxSlider.cpp and ofxInputField.cpp. Both have a function called toRange(...) which is identical in both. When the scroll callback is called this function is called. Can't we move this function to the ofxBaseGui for instance? Also, the code used in the scroll callback in both ofxSlider and ofxInputField is quite similar, maybe moving this to ofxBaseGui might be better too.