update method for drawing slider bars
Instead of calling ofPath tessellation on each draw call for sliders,
slider knows how to generate mesh for bar rectangles itself. The mesh is
cached, and only updated iff the slider value changes.

Rect mesh caching works similarly to the slider's internal caching of
text meshes.

You might notice a slight perfomance improvement on GUIs with multiple
sliders with this PR.

As an additional check/optimisation, slider fill rectangles with either
width or height < 1 pixels will also not be drawn (they would not be
visible anyway).

By *not* using ofPath (and by implication the tessellator) to draw the
slider we avoid having to ask the tessellator to tessellate extremely
small paths, in which case it had a tendency to throw its hands up in
despair or disgust (hard to tell from a stack trace) and cause crashes
(see issue #6294).

Fixes issue #6294.