Bugfix - last commit hash written before build successful. (#6355)
This isn't the cause of the current nightly build errors, but it hides the actual cause. 

Because lasthash.txt is written before the build is successful it means it's not trying to do the build each night even though the actual packages are older than the last commit.  

Moving it to the end of the script will mean that our log file will show the real error messages preventing the build.