add default member initialiser for position
+ if position was not explicitly set, this would mean that garbage
memory most likely was stored in place of the .z coordinate of a
rectangle's position.

Whilst this didn't affect any 2d operations, as soon as the rectangle in
question was going to be drawn, its .z position would be fed into the
vertex shader, and all sorts of mayhem would happen.

By setting the position to be default-initialised the .z coordinate will
receive 0, which is what any 2D graphical element would be expected to
be initialised with.