[ennhancement - openframeworks - docs] - Update README.md
[ennhancement - openframeworks - docs] - Update README.md

The commit message "[ennhancement - openframeworks - docs] - Update README.md" provides information about the nature of the changes made in the commit. Let's break it down:

[ennhancement - openframeworks - docs]:

[ennhancement]: Specifies that the changes are enhancements or improvements.
- openframeworks - docs: Indicates the context of the enhancements, specifying that they are related to the OpenFrameworks documentation.
- Update README.md:

- Update: Describes the action taken in this commit, which is an update or modification.
README.md: Specifies that the update specifically involves the README.md file.
In summary, this commit is dedicated to enhancing the README.md file within the OpenFrameworks documentation. The specifics of the enhancements made can be found by reviewing the changes within the README.md file in the "docs" directory of the OpenFrameworks project.