meson: add kissfft subproject wrap file
A wrap file basically tells meson how to download a subproject and what
dependency it fullfills. So when we do dependency('foo') in meson.build
and foo is not available in the host system, it will automatically download
it and configure it so it uses as a "subproject"

More information here https://mesonbuild.com/Wrap-dependency-system-manual.html
e.g. Archlinux doesn't bundle it